requirements.txt
streamlit
pandas
streamlit-folium
folium
google-cloud-storage
google-cloud-firestore
pillow